Build Docker image using DinD (#7)
I'm adding a new job to the pipeline that we run for every commit to ensure that the Docker image builds successfully. Unlike the release pipeline, I'm using Docker DinD. If it works, I'll update the release pipeline too to use the same approach, which is safer than exposing the host's daemon.

Reviewed-on: https://codeberg.org/glossia/glossia/pulls/7
Co-authored-by: Pedro <pedro@pepicrft.me>
Co-committed-by: Pedro <pedro@pepicrft.me>